Fix bug in Operation.finish() where passed errors were not added to _internalErrors
Previously this was combining the _internalErrors with the passed in errors
and passing them along to finished() and observer.operationDidFinish(). This
causes problems with the NoFailedDependencies condition that inspects the
_internalErrors array.

Fixed this by making finish() store the passed errors in _internalErrors.
Calling Operation.finishWithError() ends up calling finish() with the
passed error.